Fix the matching of emoji sequences
Longer sequences need to be matched first to account for sequences that
are prefixes of another one, to avoid splitting an emoji into multiple
ones.
The previous sorting by codepoints only was not handling that properly
due to basic emojis being specified as ranges for many of them, that
were sorted based on the first emoji of the range.
The new algorithm sorts first by descending length (which also avoids
reversing the array at runtime) and then by codepoints.