Reduce the size of the generated regex
Contrary to what the original library was saying, the Emoji property in
Unicode is not applied only to characters that require the emoji variant
selector. It is applied to all emoji characters. The ones that need the
variant selectors are the characters that have Emoji=Yes and
Emoji_Presentation=No.